requirements.txt
discord.py
CurrencyConverter
currency-converter
aiofiles
asyncio
aiosqlite
youtube_dl
yt_dlp